requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
chardet==3.0.4
idna==2.9
idna-ssl==1.1.0
kafka-python==2.0.1
lxml==4.5.1
multidict==4.7.6
psycopg2==2.8.5
pyspark==2.4.5
soupsieve==2.0.1
typing-extensions==3.7.4.2
yarl==1.4.2
